rename JobManagementUtil to Recorder, moved get_s3_link and get_dashboard_version to a util file
